# crates/reasondb-server/src/routes/ingest.rs
//! Ingestion endpoints
//!
//! Handle document ingestion from files, text, and URLs.

use axum::{
    extract::{Multipart, State},
    Json,
};
use reasondb_core::llm::ReasoningEngine;
use reasondb_core::text_index::TextIndex;
use reasondb_core::NodeStore;
use reasondb_ingest::{IngestPipeline, PipelineConfig};
use serde::{Deserialize, Serialize};
use std::sync::Arc;
use tempfile::NamedTempFile;
use tracing::{debug, info};
use utoipa::ToSchema;

use crate::{
    error::{ApiError, ApiResult, ErrorResponse},
    state::AppState,
};

/// Index all nodes of a document in the text index for BM25 search.
fn index_document_nodes(
    text_index: &TextIndex,
    store: &NodeStore,
    document_id: &str,
    table_id: &str,
    tags: &[String],
) -> Result<(), ApiError> {
    // Get all nodes for this document
    let nodes = store
        .get_nodes_for_document(document_id)
        .map_err(|e| ApiError::Internal(format!("Failed to get document nodes: {}", e)))?;

    // Index each node that has content
    for node in &nodes {
        // Skip nodes without content
        let content = match &node.content {
            Some(c) => c.as_str(),
            None => continue,
        };

        text_index
            .index_node(
                document_id,
                &node.id,
                table_id,
                &node.title,
                content,
                tags,
            )
            .map_err(|e| ApiError::Internal(format!("Failed to index node: {}", e)))?;
    }

    // Commit the index
    text_index
        .commit()
        .map_err(|e| ApiError::Internal(format!("Failed to commit text index: {}", e)))?;

    debug!(
        "Indexed {} nodes for document {} in BM25 index",
        nodes.len(),
        document_id
    );
    Ok(())
}

/// Response for ingestion operations
#[derive(Debug, Serialize, ToSchema)]
pub struct IngestResponse {
    /// Unique document ID
    #[schema(example = "doc_abc123")]
    pub document_id: String,
    /// Document title
    #[schema(example = "Machine Learning Fundamentals")]
    pub title: String,
    /// Total number of nodes in the tree
    #[schema(example = 42)]
    pub total_nodes: usize,
    /// Maximum depth of the tree
    #[schema(example = 4)]
    pub max_depth: usize,
    /// Ingestion statistics
    pub stats: IngestStats,
}

/// Statistics from ingestion
#[derive(Debug, Serialize, ToSchema)]
pub struct IngestStats {
    /// Characters extracted from source
    #[schema(example = 50000)]
    pub chars_extracted: usize,
    /// Text chunks created
    #[schema(example = 25)]
    pub chunks_created: usize,
    /// Tree nodes created
    #[schema(example = 42)]
    pub nodes_created: usize,
    /// Summaries generated by LLM
    #[schema(example = 42)]
    pub summaries_generated: usize,
    /// Total processing time in milliseconds
    #[schema(example = 5230)]
    pub total_time_ms: u64,
}

impl From<reasondb_ingest::IngestStats> for IngestStats {
    fn from(s: reasondb_ingest::IngestStats) -> Self {
        Self {
            chars_extracted: s.chars_extracted,
            chunks_created: s.chunks_created,
            nodes_created: s.nodes_created,
            summaries_generated: s.summaries_generated,
            total_time_ms: s.total_time_ms,
        }
    }
}

/// Request for text ingestion
#[derive(Debug, Deserialize, ToSchema)]
pub struct IngestTextRequest {
    /// Document title
    #[schema(example = "My Research Notes")]
    pub title: String,
    /// Document content (plain text or Markdown)
    #[schema(example = "# Introduction\n\nThis document covers...")]
    pub content: String,
    /// Table ID to assign the document to (REQUIRED - table must exist)
    #[schema(example = "tbl_legal")]
    pub table_id: String,
    /// Whether to generate LLM summaries (default: true)
    #[serde(default)]
    pub generate_summaries: Option<bool>,
    /// Document tags for filtering
    #[serde(default)]
    #[schema(example = json!(["nda", "confidential"]))]
    pub tags: Option<Vec<String>>,
    /// Custom metadata (key-value pairs)
    #[serde(default)]
    #[schema(example = json!({"contract_type": "nda", "value_usd": 50000}))]
    pub metadata: Option<std::collections::HashMap<String, serde_json::Value>>,
}

/// Request for URL ingestion
#[derive(Debug, Deserialize, ToSchema)]
pub struct IngestUrlRequest {
    /// URL to ingest (web page, YouTube video, etc.)
    #[schema(example = "https://en.wikipedia.org/wiki/Machine_learning")]
    pub url: String,
    /// Table ID to assign the document to (REQUIRED - table must exist)
    #[schema(example = "tbl_research")]
    pub table_id: String,
    /// Whether to generate LLM summaries (default: true)
    #[serde(default)]
    pub generate_summaries: Option<bool>,
}

/// Ingest a file (multipart upload)
///
/// Upload a document file for ingestion. Supports PDF, Word, Excel, PowerPoint,
/// images (with OCR), audio files (with transcription), HTML, and more.
#[utoipa::path(
    post,
    path = "/v1/ingest/file",
    tag = "ingestion",
    request_body(content_type = "multipart/form-data", description = "File to ingest"),
    responses(
        (status = 200, description = "File ingested successfully", body = IngestResponse),
        (status = 400, description = "Invalid request", body = ErrorResponse),
        (status = 500, description = "Ingestion failed", body = ErrorResponse),
    )
)]
pub async fn ingest_file<R: ReasoningEngine + Clone + Send + Sync + 'static>(
    State(state): State<Arc<AppState<R>>>,
    mut multipart: Multipart,
) -> ApiResult<Json<IngestResponse>> {
    // Get the file and table_id from multipart
    let mut file_data: Option<(String, Vec<u8>)> = None;
    let mut table_id: Option<String> = None;

    while let Some(field) = multipart
        .next_field()
        .await
        .map_err(|e| ApiError::BadRequest(format!("Failed to read multipart: {}", e)))?
    {
        let name = field.name().unwrap_or("").to_string();

        if name == "file" {
            let filename = field
                .file_name()
                .map(|s| s.to_string())
                .unwrap_or_else(|| "unknown".to_string());

            let data = field
                .bytes()
                .await
                .map_err(|e| ApiError::BadRequest(format!("Failed to read file: {}", e)))?;

            if data.len() > state.config.max_upload_size {
                return Err(ApiError::BadRequest(format!(
                    "File too large. Max size: {} bytes",
                    state.config.max_upload_size
                )));
            }

            file_data = Some((filename, data.to_vec()));
        } else if name == "table_id" {
            let text = field
                .text()
                .await
                .map_err(|e| ApiError::BadRequest(format!("Failed to read table_id: {}", e)))?;
            table_id = Some(text);
        }
    }

    let (filename, data) = file_data.ok_or_else(|| ApiError::BadRequest("No file provided".to_string()))?;
    let table_id = table_id.ok_or_else(|| ApiError::BadRequest("table_id is required".to_string()))?;

    // Verify the table exists
    state.store.get_table(&table_id)
        .map_err(ApiError::from)?
        .ok_or_else(|| ApiError::NotFound(format!("Table '{}' not found", table_id)))?;

    info!("Ingesting file: {} ({} bytes) into table: {}", filename, data.len(), table_id);

    // Write to temp file
    let temp_file = NamedTempFile::new()
        .map_err(|e| ApiError::Internal(format!("Failed to create temp file: {}", e)))?;

    std::fs::write(temp_file.path(), &data)
        .map_err(|e| ApiError::Internal(format!("Failed to write temp file: {}", e)))?;

    // Rename with original extension
    let extension = std::path::Path::new(&filename)
        .extension()
        .and_then(|e| e.to_str())
        .unwrap_or("bin");

    let temp_path = temp_file.path().with_extension(extension);
    std::fs::rename(temp_file.path(), &temp_path)
        .map_err(|e| ApiError::Internal(format!("Failed to rename temp file: {}", e)))?;

    // Create pipeline and ingest
    let config = PipelineConfig {
        generate_summaries: state.config.generate_summaries,
        store_in_db: true,
        ..Default::default()
    };

    let pipeline = IngestPipeline::new((*state.reasoner).clone())
        .with_config(config);

    let result = pipeline
        .ingest_and_store(&temp_path, &table_id, &state.store)
        .await
        .map_err(ApiError::from)?;

    // Clean up temp file
    let _ = std::fs::remove_file(&temp_path);

    // Index document content for BM25 search
    index_document_nodes(
        &state.text_index,
        &state.store,
        &result.document.id,
        &result.document.table_id,
        &result.document.tags,
    )?;

    debug!("Ingestion complete: {} nodes created", result.stats.nodes_created);

    Ok(Json(IngestResponse {
        document_id: result.document.id.clone(),
        title: result.document.title.clone(),
        total_nodes: result.document.total_nodes,
        max_depth: result.document.max_depth as usize,
        stats: result.stats.into(),
    }))
}

/// Ingest raw text or markdown
///
/// Ingest plain text or Markdown content directly. The content will be
/// chunked, organized into a tree structure, and optionally summarized.
/// You can optionally assign it to a table and add metadata/tags.
#[utoipa::path(
    post,
    path = "/v1/ingest/text",
    tag = "ingestion",
    request_body = IngestTextRequest,
    responses(
        (status = 200, description = "Text ingested successfully", body = IngestResponse),
        (status = 422, description = "Validation failed", body = ErrorResponse),
        (status = 500, description = "Ingestion failed", body = ErrorResponse),
    )
)]
pub async fn ingest_text<R: ReasoningEngine + Clone + Send + Sync + 'static>(
    State(state): State<Arc<AppState<R>>>,
    Json(request): Json<IngestTextRequest>,
) -> ApiResult<Json<IngestResponse>> {
    if request.title.is_empty() {
        return Err(ApiError::ValidationError("Title is required".to_string()));
    }

    if request.content.is_empty() {
        return Err(ApiError::ValidationError("Content is required".to_string()));
    }

    if request.table_id.is_empty() {
        return Err(ApiError::ValidationError("Table ID is required".to_string()));
    }

    // Verify the table exists
    state.store.get_table(&request.table_id)
        .map_err(ApiError::from)?
        .ok_or_else(|| ApiError::NotFound(format!("Table '{}' not found", request.table_id)))?;

    info!("Ingesting text: {} ({} chars) into table: {}", 
          request.title, request.content.len(), request.table_id);

    let generate_summaries = request.generate_summaries.unwrap_or(state.config.generate_summaries);

    let config = PipelineConfig {
        generate_summaries,
        store_in_db: true,
        ..Default::default()
    };

    let pipeline = IngestPipeline::new((*state.reasoner).clone())
        .with_config(config);

    let mut result = pipeline
        .ingest_text_and_store(&request.title, &request.table_id, &request.content, &state.store)
        .await
        .map_err(ApiError::from)?;

    // Apply tags and metadata if provided
    let mut doc = result.document.clone();
    let mut needs_update = false;

    if let Some(tags) = &request.tags {
        doc.tags = tags.clone();
        needs_update = true;
    }

    if let Some(metadata) = &request.metadata {
        doc.metadata = metadata.clone();
        needs_update = true;
    }

    if needs_update {
        state.store.update_document(&doc).map_err(ApiError::from)?;
        result.document = doc;
    }

    // Index document content for BM25 search
    index_document_nodes(
        &state.text_index,
        &state.store,
        &result.document.id,
        &result.document.table_id,
        &result.document.tags,
    )?;

    Ok(Json(IngestResponse {
        document_id: result.document.id.clone(),
        title: result.document.title.clone(),
        total_nodes: result.document.total_nodes,
        max_depth: result.document.max_depth as usize,
        stats: result.stats.into(),
    }))
}

/// Ingest from URL
///
/// Ingest content from a URL. Supports web pages, YouTube videos (with transcription),
/// and other online resources. The content will be extracted, chunked, and organized.
#[utoipa::path(
    post,
    path = "/v1/ingest/url",
    tag = "ingestion",
    request_body = IngestUrlRequest,
    responses(
        (status = 200, description = "URL content ingested successfully", body = IngestResponse),
        (status = 422, description = "Validation failed (invalid URL)", body = ErrorResponse),
        (status = 500, description = "Ingestion failed", body = ErrorResponse),
    )
)]
pub async fn ingest_url<R: ReasoningEngine + Clone + Send + Sync + 'static>(
    State(state): State<Arc<AppState<R>>>,
    Json(request): Json<IngestUrlRequest>,
) -> ApiResult<Json<IngestResponse>> {
    if request.url.is_empty() {
        return Err(ApiError::ValidationError("URL is required".to_string()));
    }

    if request.table_id.is_empty() {
        return Err(ApiError::ValidationError("Table ID is required".to_string()));
    }

    // Validate URL
    url::Url::parse(&request.url)
        .map_err(|e| ApiError::ValidationError(format!("Invalid URL: {}", e)))?;

    // Verify the table exists
    state.store.get_table(&request.table_id)
        .map_err(ApiError::from)?
        .ok_or_else(|| ApiError::NotFound(format!("Table '{}' not found", request.table_id)))?;

    info!("Ingesting URL: {} into table: {}", request.url, request.table_id);

    let generate_summaries = request.generate_summaries.unwrap_or(state.config.generate_summaries);

    let config = PipelineConfig {
        generate_summaries,
        store_in_db: true,
        ..Default::default()
    };

    let pipeline = IngestPipeline::new((*state.reasoner).clone())
        .with_config(config);

    let result = pipeline
        .ingest_url_and_store(&request.url, &request.table_id, &state.store)
        .await
        .map_err(ApiError::from)?;

    // Index document content for BM25 search
    index_document_nodes(
        &state.text_index,
        &state.store,
        &result.document.id,
        &result.document.table_id,
        &result.document.tags,
    )?;

    Ok(Json(IngestResponse {
        document_id: result.document.id.clone(),
        title: result.document.title.clone(),
        total_nodes: result.document.total_nodes,
        max_depth: result.document.max_depth as usize,
        stats: result.stats.into(),
    }))
}