# docs/api-reference/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReasonDB API",
    "description": "A reasoning-native database API for AI agents. ReasonDB uses LLM-guided tree traversal for intelligent document search, optimized for AI agent workflows.\n\n## Key Features\n\n- **Hierarchical Document Storage**: Documents stored as navigable trees, not flat chunks\n- **LLM-Guided Retrieval**: AI reasons through tree structure using summaries\n- **Parallel Branch Exploration**: Concurrent traversal with beam search\n- **Multi-Format Support**: PDFs, Markdown, HTML, and more\n- **RQL Query Language**: SQL-like queries with SEARCH and REASON clauses",
    "version": "0.1.0",
    "license": {
      "name": "MIT OR Apache-2.0",
      "url": "https://github.com/reasondb/reasondb"
    },
    "contact": {
      "name": "ReasonDB Contributors",
      "url": "https://github.com/reasondb/reasondb"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4444",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/v1/ingest/text": {
      "post": {
        "summary": "Ingest raw text or markdown",
        "description": "Ingest plain text or Markdown content directly. The content will be chunked, organized into a tree structure, and optionally summarized by an LLM. You must specify a table_id to assign the document to.",
        "operationId": "ingestText",
        "tags": ["ingestion"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestTextRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text ingested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Ingestion failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/file": {
      "post": {
        "summary": "Ingest a file (multipart upload)",
        "description": "Upload a document file for ingestion. Supports PDF, Word, Excel, PowerPoint, images (with OCR), audio files (with transcription), HTML, and more via MarkItDown.",
        "operationId": "ingestFile",
        "tags": ["ingestion"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file", "table_id"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File to ingest"
                  },
                  "table_id": {
                    "type": "string",
                    "description": "Table ID to assign the document to",
                    "example": "tbl_legal"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File ingested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Ingestion failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/url": {
      "post": {
        "summary": "Ingest from URL",
        "description": "Ingest content from a URL. Supports web pages, YouTube videos (with transcription), and other online resources. The content will be extracted, chunked, and organized into a tree.",
        "operationId": "ingestUrl",
        "tags": ["ingestion"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "URL content ingested successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed (invalid URL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Ingestion failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IngestTextRequest": {
        "type": "object",
        "required": ["title", "content", "table_id"],
        "properties": {
          "title": {
            "type": "string",
            "description": "Document title",
            "example": "My Research Notes"
          },
          "content": {
            "type": "string",
            "description": "Document content (plain text or Markdown)",
            "example": "# Introduction\n\nThis document covers..."
          },
          "table_id": {
            "type": "string",
            "description": "Table ID to assign the document to (REQUIRED - table must exist)",
            "example": "tbl_legal"
          },
          "generate_summaries": {
            "type": "boolean",
            "description": "Whether to generate LLM summaries (default: true)",
            "default": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Document tags for filtering",
            "example": ["nda", "confidential"]
          },
          "metadata": {
            "type": "object",
            "description": "Custom metadata (author, contract_type, etc.)",
            "additionalProperties": true,
            "example": {
              "author": "Legal Team",
              "contract_type": "nda",
              "value_usd": 50000
            }
          }
        }
      },
      "IngestUrlRequest": {
        "type": "object",
        "required": ["url", "table_id"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "URL to ingest (web page, YouTube video, etc.)",
            "example": "https://en.wikipedia.org/wiki/Machine_learning"
          },
          "table_id": {
            "type": "string",
            "description": "Table ID to assign the document to (REQUIRED - table must exist)",
            "example": "tbl_research"
          },
          "generate_summaries": {
            "type": "boolean",
            "description": "Whether to generate LLM summaries (default: true)",
            "default": true
          }
        }
      },
      "IngestResponse": {
        "type": "object",
        "properties": {
          "document_id": {
            "type": "string",
            "description": "Unique document ID",
            "example": "doc_abc123"
          },
          "title": {
            "type": "string",
            "description": "Document title",
            "example": "Machine Learning Fundamentals"
          },
          "total_nodes": {
            "type": "integer",
            "description": "Total number of nodes in the tree",
            "example": 42
          },
          "max_depth": {
            "type": "integer",
            "description": "Maximum depth of the tree",
            "example": 4
          },
          "stats": {
            "$ref": "#/components/schemas/IngestStats"
          }
        }
      },
      "IngestStats": {
        "type": "object",
        "properties": {
          "chars_extracted": {
            "type": "integer",
            "description": "Characters extracted from source",
            "example": 50000
          },
          "chunks_created": {
            "type": "integer",
            "description": "Text chunks created",
            "example": 25
          },
          "nodes_created": {
            "type": "integer",
            "description": "Tree nodes created",
            "example": 42
          },
          "summaries_generated": {
            "type": "integer",
            "description": "Summaries generated by LLM",
            "example": 42
          },
          "total_time_ms": {
            "type": "integer",
            "description": "Total processing time in milliseconds",
            "example": 5230
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "example": {
          "error": {
            "code": "NOT_FOUND",
            "message": "Resource not found: doc123"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Error code",
            "example": "NOT_FOUND",
            "enum": [
              "NOT_FOUND",
              "BAD_REQUEST",
              "VALIDATION_ERROR",
              "STORAGE_ERROR",
              "INGESTION_ERROR",
              "SEARCH_ERROR",
              "LLM_ERROR",
              "INTERNAL_ERROR"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Document not found: doc123"
          },
          "details": {
            "description": "Additional error details (optional)"
          }
        }
      }
    }
  }
}